sgr-agent-store/requirements.txt
 --extra-index-url https://erc.timetoact-group.at/
erc3>=1.0.4
openai>=2.8.1
python-dotenv
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
